requirements.txt
amqp==5.1.1
aniso8601==9.0.1
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.1
Automat==22.10.0
backcall==0.2.0
billiard==4.1.0
boto3==1.26.148
botocore==1.29.148
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmarkgfm==2022.10.27
constantly==15.1.0
cryptography==41.0.1
daphne==4.0.0
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.18
django-allauth==0.54.0
django-appconf==1.0.5
django-debug-toolbar==4.1.0
django-extensions==3.2.3
django-imagekit==4.1.0
django-multiselectfield==0.1.12
django-rest-framework==0.1.0
django-storages==1.13.2
django-taggit==4.0.0
djangorestframework==3.14.0
executing==1.2.0
graphene==3.2.2
graphene-django==3.1.1
graphql-core==3.2.3
graphql-relay==3.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipython==8.14.0
jedi==0.18.2
jmespath==1.0.1
kombu==5.3.0
Markdown==3.4.3
matplotlib-inline==0.1.6
msgpack==1.0.5
mysqlclient==2.1.1
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.5.0
promise==2.3
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7.1
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
text-unidecode==1.3
traitlets==5.9.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
zope.interface==6.0
